T285 Add a test for the constants
And reorient the tests into their own files where it matched the routes
files.
