+++ server/halomod_app/__init__.py
@@ -1,20 +1,18 @@
-from flask import Flask, jsonify, request, session, abort, send_file
+from halomod_app.routes.create import create_bp
+from halomod_app.routes.plot_types import plot_types_bp
+from halomod_app.routes.constants import constants_bp
+from halomod_app.routes.ascii import ascii_bp
+from sentry_sdk.integrations.flask import FlaskIntegration
+import sentry_sdk
+from werkzeug.exceptions import HTTPException
+from flask_session import Session
+from flask_cors import CORS
+from flask import Flask, jsonify, request, session, abort
 from . import utils
 from halomod_app.utils import get_model_names
 import base64
 import json
-import zipfile
 import dill as pickle
-from flask_cors import CORS
-import numpy as np
-from flask_session import Session
-import io
-from werkzeug.exceptions import HTTPException
-import sentry_sdk
-from sentry_sdk.integrations.flask import FlaskIntegration
-from halomod_app.routes.constants import constants_bp
-from halomod_app.routes.plot_types import plot_types_bp
-from halomod_app.routes.create import create_bp
 
 
 sess = Session()
@@ -50,6 +48,7 @@ def create_app(test_config=None):
     app.register_blueprint(constants_bp, url_prefix='/constants')
     app.register_blueprint(plot_types_bp, url_prefix='/get_plot_types')
     app.register_blueprint(create_bp, url_prefix='/create')
+    app.register_blueprint(ascii_bp, url_prefix='/ascii')
 
     # Generic Exception handler for 500 Internal Server Error
     # Returns manually formatted JSON response object with 500 code,
@@ -301,69 +300,5 @@ def create_app(test_config=None):
         response["figure"] = base64_png
 
         return jsonify(response)
-
-    # Builds and sends the text data for each model stored in the session
-    # outputs {}
-    @app.route('/ascii', methods=['GET'])
-    def ascii():
-        """ Builds and sends the text data for each model stored in the session.
-
-        get:
-          responses:
-            200:
-              description: "Returns the zip file containining the different data files for each model in the user's session"
-              content:
-                application/zip:
-        """
-        models = None
-        if 'models' in session:
-            models = pickle.loads(session.get("models"))
-        else:
-            models = {}
-
-        labels = list(models.keys())
-        objects = list(models.values())
-
-        # Open up file-like objects for response
-        buff = io.BytesIO()
-        archive = zipfile.ZipFile(buff, "w", zipfile.ZIP_DEFLATED)
-
-        # Write out mass-based, k-based and r-based data files
-        for index, object in enumerate(objects):
-            for kind in utils.XLABELS:
-
-                s = io.BytesIO()
-
-                s.write(f"# [0] {utils.XLABELS[kind]}".encode())
-
-                items = {
-                    k: utils.KEYMAP[k]["ylab"]
-                    for k in utils.KEYMAP
-                    if utils.KEYMAP[k]["xlab"] == utils.XLABELS[kind]
-                }
-
-                for j, (label, ylab) in enumerate(items.items()):
-                    if getattr(object, label) is not None:
-                        s.write(f"# [{j+1}] {ylab}".encode())
-
-                out = np.array(
-                    [getattr(object, kind)] + [
-                        getattr(object, label)
-                        for label in items
-                        if getattr(object, label) is not None
-                    ]
-                ).T
-                np.savetxt(s, out)
-
-                archive.writestr(f"{kind}Vector_{labels[index]}.txt", s.getvalue())
-
-                s.close()
-
-        archive.close()
-
-        # Reset the location of the buffer to the beginning
-        buff.seek(0)
-
-        return send_file(buff, as_attachment=True, attachment_filename="all_plots.zip")
 
     return app
